% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{app}
\alias{app}
\title{SMExplorer Shiny application object}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
app
}
\description{
SMExplorer Shiny application object
}
\keyword{datasets}

