% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{close_db_connection}
\alias{close_db_connection}
\title{Close DuckDB Connection}
\usage{
close_db_connection(con)
}
\arguments{
\item{con}{A DuckDB connection object}
}
\value{
NULL (invisibly)
}
\description{
Safely closes a DuckDB connection. Should be called when the Shiny
session ends using session$onStop().
}
\examples{
\dontrun{
session$onStop(function() {
  close_db_connection(con)
})
}

}
