% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\docType{data}
\name{consumer_key}
\alias{consumer_key}
\title{Oauth credentials}
\format{An object of class \code{character} of length 1.}
\usage{
consumer_key
}
\description{
4 character vector objects, each of length 1, containing the
OAuth credential required to download data from the Twitter API
}
\keyword{datasets}
