% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map_helpers.R
\name{create_base_map}
\alias{create_base_map}
\title{Create Base Map}
\usage{
create_base_map(spatial_data)
}
\arguments{
\item{spatial_data}{}
}
\value{
Rendered leaflet map with tiles, county polygons
}
\description{
Create base map from spatial layer to be displayed before user
 selects a metric
}
