% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWordList.R
\name{createWordList}
\alias{createWordList}
\title{Compilation of Lists on either Side of Emotive Spectrum}
\usage{
createWordList(x)
}
\arguments{
\item{x}{An atomic vector of polarities}
}
\description{
\code{createWordList} Generates list of positive and negative words.
}
