% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore Social Media Data.}
\usage{
explore()
}
\value{
This function does not return. See \code{\link{runApp}}
}
\description{
\code{explore} Launches a new Shiny application.
}
\details{
This is a wrapper function for \code{shiny::runApp}, which is a core
function for launching Shiny applications.
}
