% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{format_values}
\alias{format_values}
\title{format values
2024-11-07}
\usage{
format_values(df, units_col = "units", value_col = "value")
}
\description{
takes the value column and formats based on units to print nicely
}
