% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corpus.R
\name{make_corpus}
\alias{make_corpus}
\title{Prepare Text for Use in Sentiment Analysis}
\usage{
make_corpus(GText, stem = TRUE)
}
\arguments{
\item{GText}{A character vector}

\item{stem}{Use stemming. Defaults to \code{TRUE}}
}
\value{
A corpus
}
\description{
\code{make_corpus} Builds a corpus.
}
