% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Kernel density plots based on ggplot2 package.}
\usage{
plotDensity(data = x, entry = character(), daily = FALSE)
}
\description{
\code{plotDensity} Generates a density plot.
}

