% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareObjects.R
\name{prepareObjects}
\alias{prepareObjects}
\title{Package Objects used for Sentiment Analysis.}
\usage{
prepareObjects(data)
}
\description{
\code{prepareObjects} Internal objects used in assessing polarity.
}

