% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBagofwords.R
\name{processBagofWords}
\alias{processBagofWords}
\title{Treatment for Text ahead of Use for Plots.}
\usage{
processBagofWords(x, table)
}
\description{
\code{processBagofwords} Process text prior to visualisation.
}

