% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOnLoadedTweets.R
\docType{data}
\name{textOnLoadedTweets}
\alias{textOnLoadedTweets}
\title{Package Objects.}
\format{An object of class \code{character} of length 1.}
\usage{
textOnLoadedTweets
}
\description{
\code{TextOnLoadedTweets} A character vector for use on the user interface.
}
\keyword{datasets}

