% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOnTweetsByPlatform.R
\docType{data}
\name{textOnTweetsByPlatform}
\alias{textOnTweetsByPlatform}
\title{Package Objects.}
\format{An object of class \code{character} of length 1.}
\usage{
textOnTweetsByPlatform
}
\description{
\code{TextOnTweetsByPlatform} A string used for labelling plots.
}
\keyword{datasets}
