% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map_helpers.R
\name{create_base_map}
\alias{create_base_map}
\title{Create Base Map}
\usage{
create_base_map(spatial_data, resolution = "County")
}
\arguments{
\item{spatial_data}{Spatial data (sf object) with county or state boundaries}

\item{resolution}{Resolution level - 'County' or 'State' (default: 'County')}
}
\value{
Rendered leaflet map with tiles, polygons
}
\description{
Create base map from spatial layer to be displayed before user
 selects a metric
}
