% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map_helpers.R
\name{custom_popup}
\alias{custom_popup}
\title{Custom Popup}
\usage{
custom_popup(county_name, state_name, variable_name, value)
}
\arguments{
\item{county_name}{}

\item{state_name}{}

\item{variable_name}{}

\item{value}{}
}
\description{
Formatting for leaflet popup based on map_data inputs. Note that
this currently relies on logic in mod_map based on `input$resolution` to
account for when state data is used and there is no county_name. That should
probably be rolled into this function at some point
}
