% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map_helpers.R
\name{get_map_formulas}
\alias{get_map_formulas}
\title{Get Popup and Label Formulas}
\usage{
get_map_formulas(resolution)
}
\arguments{
\item{resolution}{Either 'County' or 'State'}
}
\value{
List with popup_formula and label_formula
}
\description{
Returns appropriate popup and label formulas based on resolution.
County data has county_name column, state data does not - this function
handles that difference.
}
