% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_welcome_helpers.R
\name{link_block}
\alias{link_block}
\title{Title}
\usage{
link_block(title = NULL, text = NULL, image = NULL, id = NULL)
}
\arguments{
\item{text}{}

\item{image}{}
}
\description{
Convenience function for formatting four link blocks on welcome page
}
