% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Kernel density plots based on ggplot2 package.}
\usage{
plotDensity(data = x, entry = character(), daily = FALSE)
}
\arguments{
\item{data}{The datafrane to be plotted from}

\item{entry}{The character input value i.e. the search term}

\item{daily}{logical - TRUE if plot represents daily data}
}
\description{
\code{plotDensity} Generates a density plot.
}

