% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareObjects.R
\name{prepareObjects}
\alias{prepareObjects}
\title{Package Objects used for Sentiment Analysis.}
\usage{
prepareObjects(data)
}
\arguments{
\item{data}{A dataframe generated from downloaded Twitter data}
}
\value{
A list of length 3 containing:
    - \code{main}: the data frame of original tweets
    (\code{isRetweet == FALSE})
    - \code{RT}: the data frame of retweets (\code{isRetweet == TRUE})
    - a vector representing the polarity of \emph{text} in \code{main}
}
\description{
\code{prepareObjects} Internal objects used in assessing polarity.
}
