% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBagofwords.R
\name{processBagofWords}
\alias{processBagofWords}
\title{Treatment for Text ahead of Use for Plots.}
\usage{
processBagofWords(x, table)
}
\arguments{
\item{x}{A data frame}

\item{table}{A list with different sign values for polarity}
}
\description{
\code{processBagofwords} Process text prior to visualisation.
}
